% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-housing.R
\name{address_lookup}
\alias{address_lookup}
\title{Address lookup service API}
\usage{
address_lookup(address, n = 1)
}
\arguments{
\item{address}{free text for address}

\item{n}{maxium number of results to return, default to 1}
}
\description{
Address lookup service API
}
