% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-health.R
\name{ae_wait_time_retrieve}
\alias{ae_wait_time_retrieve}
\title{Retrieve Data for "Hospital Authority: Accident and Emergency Waiting Time by Hospital"}
\usage{
ae_wait_time_retrieve(data_url, path = NULL)
}
\arguments{
\item{data_url}{data url as returned by `data_file_url`}

\item{path}{path to save the file}
}
\description{
Accident and emergency waiting time by hospital. \cr
\cr
UPDATE FREQUENCY: EVERY 15 MINUTES
}
