% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-health.R
\name{ae_wait_time_url}
\alias{ae_wait_time_url}
\title{Get Base URL for "Hospital Authority: Accident and Emergency Waiting Time by Hospital"}
\usage{
ae_wait_time_url()
}
\description{
Accident and emergency waiting time by hospital. \cr
\cr
UPDATE FREQUENCY: EVERY 15 MINUTES
}
