% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{aqhi_24hr}
\alias{aqhi_24hr}
\title{Retrieve Air Quality Health Index (AQHI) at individual general and roadside Air Quality Monitoring stations for the past 24 hours
Reference: https://data.gov.hk/en-data/dataset/hk-epd-airteam-past24hr-aqhi-of-individual-air-quality-monitoring-stations}
\usage{
aqhi_24hr(timestamp = NULL)
}
\arguments{
\item{timestamp}{if null then current, otherwise historical weather
it should be in format of \%Y\%m\%d-\%H\%M, e.g. 20180905-0100}
}
\description{
Retrieve Air Quality Health Index (AQHI) at individual general and roadside Air Quality Monitoring stations for the past 24 hours
Reference: https://data.gov.hk/en-data/dataset/hk-epd-airteam-past24hr-aqhi-of-individual-air-quality-monitoring-stations
}
