% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{aqhi_24hr_retrieve}
\alias{aqhi_24hr_retrieve}
\title{Retrieve Air Quality Health Index (AQHI) at individual general and roadside
Air Quality Monitoring stations for the past 24 hours}
\usage{
aqhi_24hr_retrieve(data_url)
}
\arguments{
\item{data_url}{the url to the specific file, e.g. as returned by
hist_file_url() for a given timestamp}
}
\description{
Retrieve Air Quality Health Index (AQHI) at individual general and roadside
Air Quality Monitoring stations for the past 24 hours
}
\references{
\url{https://data.gov.hk/en-data/dataset/hk-epd-airteam-past24hr-aqhi-of-individual-air-quality-monitoring-stations}
}
