% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{aqhi_current}
\alias{aqhi_current}
\title{Retrieve Current Air Quality Health Index of individual Air Quality Monitoring stations
Reference: https://data.gov.hk/en-data/dataset/hk-epd-airteam-current-aqhi-of-individual-air-quality-monitoring-stations}
\usage{
aqhi_current(timestamp = NULL)
}
\arguments{
\item{timestamp}{if null then current, otherwise historical weather
it should be in format of \%Y\%m\%d-\%H\%M, e.g. 20180905-0100}
}
\description{
Retrieve Current Air Quality Health Index of individual Air Quality Monitoring stations
Reference: https://data.gov.hk/en-data/dataset/hk-epd-airteam-current-aqhi-of-individual-air-quality-monitoring-stations
}
