% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{aqhi_current_retrieve}
\alias{aqhi_current_retrieve}
\title{Retrieve Current Air Quality Health Index of individual Air Quality Monitoring stations}
\usage{
aqhi_current_retrieve(data_url)
}
\arguments{
\item{data_url}{the url to the specific file, e.g. as returned by
hist_file_url() for a given timestamp}
}
\description{
Retrieve Current Air Quality Health Index of individual Air Quality Monitoring stations
}
\references{
\url{https://data.gov.hk/en-data/dataset/hk-epd-airteam-current-aqhi-of-individual-air-quality-monitoring-stations}
}
