% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{aqhi_current_url}
\alias{aqhi_current_url}
\title{URL for Current Air Quality Health Index of individual Air Quality Monitoring stations
Reference: https://data.gov.hk/en-data/dataset/hk-epd-airteam-current-aqhi-of-individual-air-quality-monitoring-stations}
\usage{
aqhi_current_url(lang = c("en", "tc", "sc"))
}
\arguments{
\item{lang}{select the language of the available data, available choices are
"en" for English, "tc" for traditional Chinese, "sc" for simplified Chinese}
}
\description{
URL for Current Air Quality Health Index of individual Air Quality Monitoring stations
Reference: https://data.gov.hk/en-data/dataset/hk-epd-airteam-current-aqhi-of-individual-air-quality-monitoring-stations
}
