% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{aqhi_range_forecast}
\alias{aqhi_range_forecast}
\title{Retrieve Current Air Quality Health Index Range and Forecast
Reference: https://data.gov.hk/en-data/dataset/hk-epd-airteam-current-air-quality-health-index-range-and-forecast}
\usage{
aqhi_range_forecast(timestamp = NULL)
}
\arguments{
\item{timestamp}{if null then current, otherwise historical weather
it should be in format of \%Y\%m\%d-\%H\%M, e.g. 20180905-0100}
}
\description{
Retrieve Current Air Quality Health Index Range and Forecast
Reference: https://data.gov.hk/en-data/dataset/hk-epd-airteam-current-air-quality-health-index-range-and-forecast
}
