% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{aqhi_range_forecast_retrieve}
\alias{aqhi_range_forecast_retrieve}
\title{Retrieve Current Air Quality Health Index Range and Forecast}
\usage{
aqhi_range_forecast_retrieve(data_url)
}
\arguments{
\item{data_url}{the url to the specific file, e.g. as returned by
hist_file_url() for a given timestamp}
}
\description{
Retrieve Current Air Quality Health Index Range and Forecast
}
\references{
\url{https://data.gov.hk/en-data/dataset/hk-epd-airteam-current-air-quality-health-index-range-and-forecast}
}
