% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{aqhi_range_forecast_url}
\alias{aqhi_range_forecast_url}
\title{URL for Current Air Quality Health Index Range and Forecast
Reference: https://data.gov.hk/en-data/dataset/hk-epd-airteam-current-air-quality-health-index-range-and-forecast}
\usage{
aqhi_range_forecast_url(lang = c("en", "tc", "sc"))
}
\arguments{
\item{lang}{select the language of the available data, available choices are
"en" for English, "tc" for traditional Chinese, "sc" for simplified Chinese}
}
\description{
URL for Current Air Quality Health Index Range and Forecast
Reference: https://data.gov.hk/en-data/dataset/hk-epd-airteam-current-air-quality-health-index-range-and-forecast
}
