% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-timestamp.R
\name{data_file_url}
\alias{data_file_url}
\title{Historical Archive File URL For A Specific Version}
\usage{
data_file_url(url, timestamp = NULL)
}
\arguments{
\item{url}{url of the historical files}

\item{timestamp}{timestamp of the historical file to retrieve. if null then
current, otherwise historical it should be in format of \%Y\%m\%d-\%H\%M,
e.g. 20180905-1306}
}
\description{
Historical Archive File URL For A Specific Version
}
