% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-health.R
\name{ev_surveillance_retrieve}
\alias{ev_surveillance_retrieve}
\title{Retrieve Data for "Department of Health: EV Scan's Figures Data"}
\usage{
ev_surveillance_retrieve(url, path = ".", keep = FALSE)
}
\arguments{
\item{url}{base url as returned by `ev_surveillance_url`}

\item{path}{path to save the file}

\item{keep}{whether to keep the file after read}
}
\description{
Hand, foot and mouth disease surveillance data including number of EV71 cases, 
institutional outbreaks, hospital surveillance and sentinel surveillance. \cr
\cr
UPDATE FREQUENCY: WEEKLY
}
