% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-health.R
\name{ev_surveillance_url}
\alias{ev_surveillance_url}
\title{Get Base URL for "Department of Health: EV Scan's Figures Data"}
\usage{
ev_surveillance_url()
}
\description{
Hand, foot and mouth disease surveillance data including number of EV71 cases, 
institutional outbreaks, hospital surveillance and sentinel surveillance. \cr
\cr
UPDATE FREQUENCY: WEEKLY
}
