% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-health.R
\name{flu_surveillance_retrieve}
\alias{flu_surveillance_retrieve}
\title{Retrieve Data for "Department of Health: Flu Express's Figures Data"}
\usage{
flu_surveillance_retrieve(url, path = ".", keep = FALSE)
}
\arguments{
\item{url}{base url as returned by `flu_surveillance_url`}

\item{path}{path to save the file}

\item{keep}{whether to keep the file after read}
}
\description{
Influenza surveillance data including sentinel surveillance, laboratory surveillance, influenza-like illness outbreak, hospital surveillance and severe influenza case during influenza season.
\cr
UPDATE FREQUENCY: WEEKLY
}
