% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-health.R
\name{flu_surveillance_url}
\alias{flu_surveillance_url}
\title{Get Base URL for "Department of Health: Flu Express's Figures Data"}
\usage{
flu_surveillance_url()
}
\description{
Influenza surveillance data including sentinel surveillance, laboratory surveillance, influenza-like illness outbreak, hospital surveillance and severe influenza case during influenza season.
\cr
UPDATE FREQUENCY: WEEKLY
}
