% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-file.R
\name{get_file_csv}
\alias{get_file_csv}
\title{Get csv format file}
\usage{
get_file_csv(url, path = NULL)
}
\arguments{
\item{url}{the url of the file}

\item{path}{the directory where the file should be save, if NULL it
will not save}
}
\value{
it returns a data frame read from the csv file
}
\description{
Get csv format file
}
\examples{
data <- get_file_csv("http://www.chp.gov.hk/files/misc/statistics-2017-vss-eng.csv")

}
