% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-file.R
\name{get_file_jpg}
\alias{get_file_jpg}
\title{Get jpg format file}
\usage{
get_file_jpg(url, path = ".", view = TRUE)
}
\arguments{
\item{url}{the url of the file}

\item{path}{the directory to save the file}

\item{view}{if TRUE the the file is viewed after download}
}
\description{
Get jpg format file
}
\examples{
get_file_jpg("http://www.toothclub.gov.hk/chi/jpg/psi/P041.jpg")

}
