% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-file.R
\name{get_file_json}
\alias{get_file_json}
\title{Get json format file}
\usage{
get_file_json(url, path = NULL)
}
\arguments{
\item{url}{the url of the file}

\item{path}{the directory where the file should be save, if NULL it
will not save}
}
\value{
it returns a data read from the json file
}
\description{
Get json format file
}
\examples{
data <- get_file_csv("http://www.ha.org.hk/opendata/ipdpdd-en.json")

}
