% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-timestamp.R
\name{get_file_versions}
\alias{get_file_versions}
\title{Retrieve Versions of Historical Archive Files}
\usage{
get_file_versions(url, start, end = NULL, retrieve_func, ...)
}
\arguments{
\item{url}{url of the file}

\item{start}{start date}

\item{end}{end date}

\item{...}{extra parameters to pass to retireve_func()}

\item{retireve_func}{function that when given a file url (obtained from
timestamp), would retrieve the specified data}
}
\description{
Retrieve Versions of Historical Archive Files
}
