% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-file.R
\name{get_file_xlsx}
\alias{get_file_xlsx}
\title{Get xlsx format file}
\usage{
get_file_xlsx(url, path = ".", silent = FALSE)
}
\arguments{
\item{url}{the url of the file}

\item{path}{the directory where the file should be downloaded to}

\item{silient}{whether or not silent message}
}
\description{
Get xlsx format file
}
