% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-list.R
\name{hist_file_list}
\alias{hist_file_list}
\title{Historical Archive File list}
\usage{
hist_file_list(start, end, category = NULL, provider = NULL,
  format = NULL, search = NULL, order = NULL, skip = NULL)
}
\arguments{
\item{start}{start date of a period}

\item{end}{end date of a period}

\item{category}{Category ID, see website for a list of categories}

\item{provider}{Provider ID, see website for a list of provider ID}

\item{format}{File format}

\item{search}{keyword search}

\item{order}{sort order}

\item{skip}{the first x number of records to omit}
}
\description{
Historical Archive File list
}
\examples{
hist_file_list("2015-01-01", "2018-01-01")

hist_file_list("2015-01-01", "2018-01-01", category = "health")

}
