% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-file.R
\name{hist_file_url}
\alias{hist_file_url}
\title{Historical Archive File URL For A Specific Version}
\usage{
hist_file_url(url, timestamp)
}
\arguments{
\item{url}{url of the historical files}

\item{timestamp}{timestamp of the historical file to retrieve}
}
\description{
Historical Archive File URL For A Specific Version
}
