% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-census.R
\name{hk_census_stpug}
\alias{hk_census_stpug}
\title{HK 2006 Population Census: by Small Tertiary Planning Unit Group (STPUG)}
\usage{
hk_census_stpug(table_id = NULL, keyword = NULL)
}
\arguments{
\item{keyword}{the keyword/category of the table}

\item{table_no}{the number of the table}
}
\description{
HK 2006 Population Census: by Small Tertiary Planning Unit Group (STPUG)
}
\details{
totally got 11 cencus statistics tables at STPUG level: \cr
1) you can search by 'Table_ID' 
No Table_ID Table_Keywords Table_Description \cr
1	A501  Hong Kong Resident Population by Ethnicity and Small Tertiary Planning Unit Group, 2006 (A501) [English] \cr
2	A502	Hong Kong Resident Population Aged 5 and Over by Usual Language and Small Tertiary Planning Unit Group, 2006 (A502) [English] \cr
3	A503	Hong Kong Resident Population by Broad Age Group, Sex and Small Tertiary Planning Unit Group, 2006 (A503) [English] \cr
4	C501	Hong Kong Resident Population by Economic Activity Status and Small Tertiary Planning Unit Group, 2006 (C501) [English] \cr
5	C502	Working Population by Industry (Sector) and Small Tertiary Planning Unit Group, 2006 (C502) [English] \cr
6	C503	Working Population by Occupation (Major Group) and Small Tertiary Planning Unit Group, 2006 (C503) [English] \cr
7	C504	Working Population by Monthly Income from Main Employment and Small Tertiary Planning Unit Group, 2006 (C504) [English] \cr
8	D401	Domestic Households by Household Size and Small Tertiary Planning Unit Group, 2006 (D401) [English] \cr
9	D402	Domestic Households by Household Composition and Small Tertiary Planning Unit Group, 2006 (D402) [English] \cr
10	D403	Domestic Households by Monthly Domestic Household Income and Small Tertiary Planning Unit Group, 2006 (D403) [English] \cr
11	E501	Occupied Quarters (Land) by Type of Quarters and Small Tertiary Planning Unit Group, 2006 (E501) [English] \cr
\cr
2) or search by keywords: \cr
ethnicity, language, age, sex, economic_activity_status \cr
industry, occupation, monthly_income \cr
household_income, household_size, household_composition, quarter_type
}
\examples{
 
hk_census_stpug(table_id = "A501")
hk_census_stpug(keyword = "household_income")
}
