% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{hk_holidays}
\alias{hk_holidays}
\title{Retrieve Hong Kong Public Holidays Data}
\usage{
hk_holidays(path = NULL)
}
\description{
Reference: https://data.gov.hk/en-data/dataset/hk-effo-statistic-cal
1823 currently provides data of Hong Kong Public Holidays for 2018-2020.
}
\details{
#' @param path the directory where the raw file should be save, if NULL it
will not be saved
}
