% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-ia.R
\name{hkia_l1}
\alias{hkia_l1}
\title{HKIA: Market Statistics - L1 Individual Life Business}
\usage{
hkia_l1(year, path = ".", keep = FALSE)
}
\arguments{
\item{year}{year of the data}

\item{path}{path to save the file}

\item{keep}{whether to keep the file after read}
}
\description{
HKIA: Market Statistics - L1 Individual Life Business
}
