% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-ia.R
\name{hkia_l13}
\alias{hkia_l13}
\title{HKIA: Individual Insurers' Statistics - L13 Total In-Force Business}
\usage{
hkia_l13(year, path = ".", keep = FALSE)
}
\arguments{
\item{year}{year of the data}

\item{path}{path to save the file}

\item{keep}{whether to keep the file after read}
}
\description{
HKIA: Individual Insurers' Statistics - L13 Total In-Force Business
}
