% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-ia.R
\name{hkia_long_term_provisional}
\alias{hkia_long_term_provisional}
\title{Quarterly Release of Provisional Statistics for Long Term Business.
Results are cumulative}
\usage{
hkia_long_term_provisional(year, quarter, path = ".", keep = FALSE)
}
\arguments{
\item{year}{the year}

\item{quarter}{the quarter}

\item{path}{path to save the file}

\item{keep}{whether to keep the file after read}
}
\description{
Quarterly Release of Provisional Statistics for Long Term Business.
Results are cumulative
}
\details{
Currently we only parse sheets for L1s (new direct individual business) and 
L3s (inforce direct individual business)
}
