% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-ia.R
\name{hkia_longterm_insurer}
\alias{hkia_longterm_insurer}
\title{HKIA: Annual Statistics for Long Term Business - Individual Insurers}
\usage{
hkia_longterm_insurer(year, table, path = ".", keep = FALSE)
}
\arguments{
\item{year}{year of the data}

\item{table}{the table name L8 - L19}

\item{path}{path to save the file}

\item{keep}{whether to keep the file after read}
}
\description{
HKIA: Annual Statistics for Long Term Business - Individual Insurers
}
\details{
The list of available tables:\cr
L8 Non-Linked Individual Life In-Force Business \cr
L9 Linked Individual Life In-Force Business\cr
L10 Group Life In-Force Business\cr
L11 Retirement Scheme In-Force Business \cr
L12 Annuity and Other In-Force Business  \cr
L13 Total In-Force Business \cr
L14 Non-Linked Individual Life New Business\cr
L15 Linked Individual Life New Business \cr
L16 Total Individual Life New Business \cr
L17 Non-Linked Individual Annuity New Business\cr
L18 Linked Individual Annuity New Business\cr
L19 Total Individual Annuity New Business \cr
}
