% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-ia.R
\name{hkia_longterm_market}
\alias{hkia_longterm_market}
\title{HKIA: Market Statistics}
\usage{
hkia_longterm_market(year, table, path = ".", keep = FALSE)
}
\arguments{
\item{year}{year of the data}

\item{table}{the table name L1 - L7}

\item{path}{path to save the file}

\item{keep}{whether to keep the file after read}
}
\description{
HKIA: Market Statistics
}
\details{
The list of available tables:\cr
L1 Individual Life Business\cr
L2 Non-Linked Individual Life In-Force Business \cr
L3 Linked Individual Life In-Force Business\cr
L4 Individual Life New Business\cr
L6 Group Life and Retirement Scheme In-Force Business\cr
L7 Annuity and Other In-Force Business/Individual Annuity New Business\cr
L8 Non-Linked Individual Life In-Force Business
}
