% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-health.R
\name{infectious_diseases_retrieve}
\alias{infectious_diseases_retrieve}
\title{Retrieve Data for "Department of Health: Number of Notifiable Infectious Diseases by Month Share"}
\usage{
infectious_diseases_retrieve(url, path = NULL)
}
\arguments{
\item{url}{base url as returned by `infectious_diseases_url`}

\item{path}{path to save the file}
}
\description{
Number of notifiable infectious diseases by month. \cr
\cr
UPDATE FREQUENCY: MONTHLY
}
