% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-health.R
\name{infectious_diseases_url}
\alias{infectious_diseases_url}
\title{Get Base URL for "Department of Health: Number of Notifiable Infectious Diseases by Month Share"}
\usage{
infectious_diseases_url(year)
}
\arguments{
\item{year}{year of the data (available since 1997)}
}
\description{
Number of notifiable infectious diseases by month. \cr
\cr
UPDATE FREQUENCY: MONTHLY
}
