% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-health.R
\name{inpatient_statistics}
\alias{inpatient_statistics}
\title{Get Data for "Department of Health: Inpatient Statistics"}
\format{A data frame with 10 variables.\cr
* `icd_code`: Detailed List No. ICD 10th Revision\cr
* `dx_grp`: Disease group\cr
* `ip_ha_hosp`: Number of inpatient discharges and deaths - Hospital Authority hospitals\cr
* `ip_ci_hosp`: Number of inpatient discharges and deaths - Correctional Institution hospitals\cr
* `ip_pvt_hosp`: Number of inpatient discharges and deaths - Private hospitals\cr
* `ip_total`: Number of inpatient discharges and deaths - Total\cr
* `reg_dealth_male`: Registered deaths in Hong Kong - Male\cr
* `reg_dealth_female`: Registered deaths in Hong Kong - Female\cr
* `reg_dealth_total`: Registered deaths in Hong Kong - Total\cr
* `reg_dealth_unknown`: Registered deaths in Hong Kong - Unknown sex}
\source{
<https://data.gov.hk/en-data/dataset/hk-dh-dh_ncddhss-ncdd-dataset-10>
}
\usage{
inpatient_statistics(year, path = ".", keep = FALSE)
}
\arguments{
\item{year}{year of the data (available since 2012)}

\item{path}{path to save the file}

\item{keep}{whether to keep the file after read}
}
\description{
Inpatient discharges and deaths in all hospitals classified by disease. \cr
\cr
UPDATE FREQUENCY: AS AND WHEN NECESSARY
}
\details{
* Classification of diseases and causes of death is based on the International\cr
Statistical Classification of Diseases and Related Health Problems (ICD) 10th\cr
Revision from 2001 onwards. Figures from 2001 onwards may not be comparable\cr
with figures for previous years which were compiled based on the ICD 9th Revision.\cr
* Chinese terms of disease groups are based on those compiled by WHO Collaborating Center for the Classification of Diseases, Beijing as the benchmark.\cr
* Refers to discharges and deaths on attendance basis.\cr
Patients treated in nursing homes are excluded.\cr
* The number of inpatient discharges and deaths for Hospital Authority hospitals\cr
may be slightly different from that presented in "Tables on Health Status and\cr
Health Services 2012" Table A18 due to different sources of data.
}
