% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-health.R
\name{inpatient_statistics_retrieve}
\alias{inpatient_statistics_retrieve}
\title{Retrieve Data for "Department of Health: Inpatient Statistics"}
\usage{
inpatient_statistics_retrieve(url, path = ".", keep = FALSE)
}
\arguments{
\item{url}{base url as returned by `inpatient_statistics_url`}

\item{path}{path to save the file}

\item{keep}{whether to keep the file after read}
}
\description{
Inpatient discharges and deaths in all hospitals classified by disease. \cr
\cr
UPDATE FREQUENCY: AS AND WHEN NECESSARY
}
