% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-health.R
\name{inpatient_statistics_url}
\alias{inpatient_statistics_url}
\title{Get Base URL for "Department of Health: Inpatient Statistics"}
\usage{
inpatient_statistics_url(year)
}
\arguments{
\item{year}{year of the data (available since 2012)}
}
\description{
Inpatient discharges and deaths in all hospitals classified by disease. \cr
\cr
UPDATE FREQUENCY: AS AND WHEN NECESSARY
}
