% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{lamp_posts_data}
\alias{lamp_posts_data}
\title{Retrieve Real-time city data collected by multi-purpose lamp posts in Kowloon East}
\usage{
lamp_posts_data(path = NULL)
}
\description{
Reference: https://data.gov.hk/en-data/dataset/hk-devb-mplp-mplp-sensor-data
Note that only the most update value is available.
}
\details{
#' @param path the directory where the raw file should be save, if NULL it
will not be saved
}
