% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{lamp_posts_data_url}
\alias{lamp_posts_data_url}
\title{URL for Real-time city data collected by multi-purpose lamp posts in Kowloon East}
\usage{
lamp_posts_data_url()
}
\description{
Reference: https://data.gov.hk/en-data/dataset/hk-devb-mplp-mplp-sensor-data
Note that only the most update value is available.
}
