% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-file.R
\name{list_hist_file_nomax}
\alias{list_hist_file_nomax}
\title{List all historical files without maximum}
\usage{
list_hist_file_nomax(...)
}
\arguments{
\item{...}{parameters passed to list_hist_file function}
}
\description{
It continuesly call list_hist_file until no result returned
}
