% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-health.R
\name{notifiable_infectious_diseases}
\alias{notifiable_infectious_diseases}
\title{Department of Health: Number of Notifiable Infectious Diseases by Month Share}
\format{A data frame with 14 variables.
* `disease`: Disease name
* `jan`-`dec`: Number of cases in each month
* `total`: Total number of cases in the year}
\source{
<https://data.gov.hk/en-data/dataset/hk-dh-chpsebcdd-number-of-notifications-for-notifiable-infectious-diseases>
}
\usage{
notifiable_infectious_diseases(year, path = ".")
}
\arguments{
\item{year}{year of the data (available since 1997)}

\item{path}{path to save the file}
}
\description{
Number of notifiable infectious diseases by month \cr
\cr
UPDATE FREQUENCY: MONTHLY
}
