% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{past_aqhi}
\alias{past_aqhi}
\title{Retrieve Past record of Air Quality Health Index}
\usage{
past_aqhi(year, month, path = NULL)
}
\arguments{
\item{year}{year of the past record}

\item{month}{month of the past record}

\item{path}{the directory where the raw file should be save, if NULL it
will not be saved}
}
\description{
Reference: https://data.gov.hk/en-data/dataset/hk-epd-airteam-past-record-of-air-quality-health-index-en
Only available from Decemenber 2013 onward.
}
