% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{past_aqhi_url}
\alias{past_aqhi_url}
\title{URL for Past record of Air Quality Health Index}
\usage{
past_aqhi_url(year, month)
}
\arguments{
\item{year}{year of the past record}

\item{month}{month of the past record}
}
\description{
Reference: https://data.gov.hk/en-data/dataset/hk-epd-airteam-past-record-of-air-quality-health-index-en
Only available from Decemenber 2013 onward.
}
