% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{past_pollution_index}
\alias{past_pollution_index}
\title{Retrieve Past hourly record of Air Pollution Index at a given year and month}
\usage{
past_pollution_index(year, month, path = NULL)
}
\arguments{
\item{year}{year of the past record}

\item{month}{month of the past record}

\item{path}{the directory where the raw file should be save, if NULL it
will not be saved}
}
\description{
Reference: https://data.gov.hk/en-data/dataset/hk-epd-airteam-past-record-of-air-pollution-index-en
Only available from July 1999 to Decemenber 2013.
}
