% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{past_pollution_index_url}
\alias{past_pollution_index_url}
\title{URL for Past hourly record of Air Pollution Index at a given year and month}
\usage{
past_pollution_index_url(year, month)
}
\arguments{
\item{year}{year of the past record}

\item{month}{month of the past record}
}
\description{
Reference: https://data.gov.hk/en-data/dataset/hk-epd-airteam-past-record-of-air-pollution-index-en
Only available from July 1999 to Decemenber 2013.
}
