% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{pollutant_24hr_retrieve}
\alias{pollutant_24hr_retrieve}
\title{Retrieve Past 24-hour Pollutant Concentration of individual Air Quality Monitoring stations}
\usage{
pollutant_24hr_retrieve(data_url)
}
\arguments{
\item{data_url}{the url to the specific file, e.g. as returned by
hist_file_url() for a given timestamp}
}
\description{
Retrieve Past 24-hour Pollutant Concentration of individual Air Quality Monitoring stations
}
\references{
\url{https://data.gov.hk/en-data/dataset/hk-epd-airteam-past24hr-pc-of-individual-air-quality-monitoring-stations}
}
