% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{pollutant_24hr_url}
\alias{pollutant_24hr_url}
\title{URL for Past 24-hour Pollutant Concentration of individual Air Quality Monitoring stations}
\usage{
pollutant_24hr_url(lang = "en")
}
\arguments{
\item{lang}{select the language of the available data, available choices are
"en" for English}
}
\description{
URL for Past 24-hour Pollutant Concentration of individual Air Quality Monitoring stations
}
\references{
\url{https://data.gov.hk/en-data/dataset/hk-epd-airteam-past24hr-pc-of-individual-air-quality-monitoring-stations}
}
