% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-housing.R
\name{public_housing_estates}
\alias{public_housing_estates}
\title{Location and Profile of Public housing estates (Annually update)}
\usage{
public_housing_estates(lang = "en")
}
\arguments{
\item{lang}{the language: en, tc, sc}
}
\description{
Location and Profile of Public housing estates (Annually update)
}
