% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-housing.R
\name{public_housing_rental}
\alias{public_housing_rental}
\title{Housing Authority's Public Rental Housing Stock (Quarterly update)}
\usage{
public_housing_rental()
}
\description{
Housing Authority's Public Rental Housing Stock (Quarterly update)
}
