% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-housing.R
\name{public_housing_trans}
\alias{public_housing_trans}
\title{Transactions in the Sale of Home Ownership Scheme (HOS) Flats and 
Transactions in the Sale of Green Form Subsidised Home Ownership Scheme (GSH) 
Flats}
\usage{
public_housing_trans(year = 2018)
}
\arguments{
\item{year}{the year of data currently only 2018, will adjust when more data
available}
}
\description{
Transactions in the Sale of Home Ownership Scheme (HOS) Flats and 
Transactions in the Sale of Green Form Subsidised Home Ownership Scheme (GSH) 
Flats
}
