% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{towngas_performance_data}
\alias{towngas_performance_data}
\title{Retrieve Towngas Environmental Performance Data}
\usage{
towngas_performance_data(fromYear, toYear = fromYear, path = NULL)
}
\arguments{
\item{fromYear}{integer, the starting year of the performance data.}

\item{toYear}{integer, the ending (inclusive) year of the performance data.}

\item{path}{the directory where the raw file should be save, if NULL it
will not be saved}
}
\description{
Retrieve Towngas Environmental Performance Data
}
\references{
\url{https://data.gov.hk/en-data/dataset/towngas-towngas-environment}
}
