% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{towngas_performance_data_url}
\alias{towngas_performance_data_url}
\title{URL for Towngas Environmental Performance Data}
\usage{
towngas_performance_data_url(fromYear, toYear = fromYear)
}
\arguments{
\item{fromYear}{integer, the starting year of the performance data.}

\item{toYear}{integer, the ending (inclusive) year of the performance data.}
}
\description{
URL for Towngas Environmental Performance Data
}
\references{
\url{https://data.gov.hk/en-data/dataset/towngas-towngas-environment}
}
