% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-transport.R
\name{transport_Arrival_CrossBoundaryFerry}
\alias{transport_Arrival_CrossBoundaryFerry}
\title{Retrieve the Arrival Information for Cross Boundary Ferry Service given specific timestamp and language version}
\usage{
transport_Arrival_CrossBoundaryFerry(timestamp = NULL, Lang = "en")
}
\arguments{
\item{timestamp}{Timestamp of the historical file to retrieve. If null then
current, otherwise historical. It should be in format of \%Y\%m\%d-\%H\%M,
e.g. 20180905-1306.}

\item{Lang}{Options of Language: "sc": simplified Chinese; "tc": traditional Chinese; "en": English (default)}
}
\description{
Retrieve the Arrival Information for Cross Boundary Ferry Service given specific timestamp and language version
}
