% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-transport.R
\name{transport_Arrival_CrossBoundaryFerry_retrieve}
\alias{transport_Arrival_CrossBoundaryFerry_retrieve}
\title{Retrieve the Arrival Information for Cross Boundary Ferry Service given specific data url}
\usage{
transport_Arrival_CrossBoundaryFerry_retrieve(data_url)
}
\arguments{
\item{data_url}{Specific data url}
}
\description{
Retrieve the Arrival Information for Cross Boundary Ferry Service given specific data url
}
