% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-transport.R
\name{transport_Arrival_CrossBoundaryFerry_url}
\alias{transport_Arrival_CrossBoundaryFerry_url}
\title{Generate the basic url for Arrival Information for Cross Boundary Ferry Service}
\usage{
transport_Arrival_CrossBoundaryFerry_url(Lang = "en")
}
\arguments{
\item{Lang}{Options of Language: "sc": simplified Chinese; "tc": traditional Chinese; "en": English (default)}
}
\description{
Generate the basic url for Arrival Information for Cross Boundary Ferry Service
}
