% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-transport.R
\name{transport_carpark_basic_retrieve}
\alias{transport_carpark_basic_retrieve}
\title{Retrieve the Basic Information of Participating Car Parks given specific data url}
\usage{
transport_carpark_basic_retrieve(data_url)
}
\arguments{
\item{data_url}{Specific data url}
}
\description{
Retrieve the Basic Information of Participating Car Parks given specific data url
}
