% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-transport.R
\name{transport_carpark_vacancy}
\alias{transport_carpark_vacancy}
\title{Retrieve the Parking Vacancy Data of Participating Car Parks given specific timestamp}
\usage{
transport_carpark_vacancy(timestamp = NULL)
}
\arguments{
\item{timestamp}{Timestamp of the historical file to retrieve. If null then
current, otherwise historical. It should be in format of \%Y\%m\%d-\%H\%M,
e.g. 20180905-1306.}
}
\description{
Retrieve the Parking Vacancy Data of Participating Car Parks given specific timestamp
}
