% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-transport.R
\name{transport_carpark_vacancy_retrieve}
\alias{transport_carpark_vacancy_retrieve}
\title{Retrieve the Parking Vacancy Data of Participating Car Parks given specific data url}
\usage{
transport_carpark_vacancy_retrieve(data_url)
}
\arguments{
\item{data_url}{Specific data url}
}
\description{
Retrieve the Parking Vacancy Data of Participating Car Parks given specific data url
}
