% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-transport.R
\name{transport_carpark_vacancy_url}
\alias{transport_carpark_vacancy_url}
\title{Generate the basic url for Parking Vacancy Data of Participating Car Parks}
\usage{
transport_carpark_vacancy_url()
}
\description{
Generate the basic url for Parking Vacancy Data of Participating Car Parks
}
