% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-transport.R
\name{transport_EV_charging}
\alias{transport_EV_charging}
\title{Retrieve details of nearby Electric EV charging stations}
\usage{
transport_EV_charging(lat_in, long_in)
}
\arguments{
\item{lat_in}{Input the latitude of the current place}

\item{long_in}{Input the longitude of the current place}
}
\description{
Retrieve details of nearby Electric EV charging stations
}
