% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-transport.R
\name{transport_Flight}
\alias{transport_Flight}
\title{Generate the list of Flight Information given the time range, arrival or departure, cargo or passenger for different language versions}
\usage{
transport_Flight(start, end, Arrival, Cargo, Lang = "en")
}
\arguments{
\item{start}{String, "YYYY-MM-DD"  starting date of interest}

\item{end}{String, "YYYY-MM-DD"  ending date of interest}

\item{Arrival}{Arrival or departure (TRUE=arrival, FALSE=departure)}

\item{Cargo}{Cargo or passenger flight (TRUE=cargo, FALSE=passenger)}

\item{Lang}{Options of Language: "zh_HK": traditional Chinese; "zh_CN": simplified Chinese; "en": English (default)}
}
\description{
Generate the list of Flight Information given the time range, arrival or departure, cargo or passenger for different language versions
}
