% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-weather.R
\name{weather_forecast_9day}
\alias{weather_forecast_9day}
\title{Retrieve current and historical 9day weather forecast}
\usage{
weather_forecast_9day(timestamp = NULL)
}
\arguments{
\item{timestamp}{if null then current, otherwise historical weather
it should be in format of \%Y\%m\%d-\%H\%M, e.g. 20180905-0100}
}
\description{
Retrieve current and historical 9day weather forecast
}
