% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{aqhi_current}
\alias{aqhi_current}
\title{Retrieve Current Air Quality Health Index of individual Air Quality Monitoring stations}
\usage{
aqhi_current(timestamp = NULL, lang = c("en", "tc", "sc"))
}
\arguments{
\item{timestamp}{if null then current, otherwise historical weather
it should be in format of \%Y\%m\%d-\%H\%M, e.g. 20180905-0100}

\item{lang}{select the language of the available data, available choices are
"en" for English, "tc" for traditional Chinese, "sc" for simplified Chinese}
}
\description{
Retrieve Current Air Quality Health Index of individual Air Quality Monitoring stations
}
\references{
\url{https://data.gov.hk/en-data/dataset/hk-epd-airteam-current-aqhi-of-individual-air-quality-monitoring-stations}
}
