% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-pollution.R
\name{aqhi_range_forecast}
\alias{aqhi_range_forecast}
\title{Retrieve Current Air Quality Health Index Range and Forecast}
\usage{
aqhi_range_forecast(timestamp = NULL, lang = c("en", "tc", "sc"))
}
\arguments{
\item{timestamp}{if null then current, otherwise historical weather
it should be in format of \%Y\%m\%d-\%H\%M, e.g. 20180905-0100}

\item{lang}{select the language of the available data, available choices are
"en" for English, "tc" for traditional Chinese, "sc" for simplified Chinese}
}
\description{
Retrieve Current Air Quality Health Index Range and Forecast
}
\references{
\url{https://data.gov.hk/en-data/dataset/hk-epd-airteam-current-air-quality-health-index-range-and-forecast}
}
