% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-holidays.R
\name{hk_holidays}
\alias{hk_holidays}
\title{Retrieve Hong Kong Public Holidays Data}
\usage{
hk_holidays(path = NULL, lang = c("en", "tc", "sc"))
}
\arguments{
\item{path}{the directory where the raw file should be save, if NULL it
will not be saved}

\item{lang}{select the language of the available data, available choices are
"en" for English, "tc" for traditional Chinese, "sc" for simplified Chinese}
}
\description{
Retrieve Hong Kong Public Holidays Data
}
\references{
\url{https://data.gov.hk/en-data/dataset/hk-effo-statistic-cal}
1823 currently provides data of Hong Kong Public Holidays for 2018-2020.
}
